Refactor: Category, ProgressStatus 색상 컴포넌트화#61
