+++ src/app/(admin)/admin/study/[id]/projectDetailPage.tsx
@@ -9,11 +9,12 @@ import { getProjectMaterials } from "@/mocks/mockProjectData";
 import BackToListButton from "@/components/detail/SCBackToListButton";
 import KebabMenu from "@/components/detail/CCKebabMenu";
 import ShareButton from "@/components/detail/CCShareButton";
-import { getStatusColor } from "@/utils/projectUtils";
-import { STATUS_LABELS, AUTHOR_STATUS_LABELS } from "@/types/project";
+import { AUTHOR_STATUS_LABELS } from "@/types/project";
 import DefaultBadge from "@/components/ui/defaultBadge";
 import MeetingMinutes from "@/components/study/CCMeetingMinutes";
 import EndRequestButton from "@/components/ui/endRequestButton";
+import { STATUS_LABELS } from "@/types/progressStatus";
+import { getStatusColor } from "@/utils/badgeUtils";
 
 interface ProjectDetailPageProps {
   params: { id: string };
