+++ src/app/(admin)/admin/study/[id]/studyDetailPage.tsx
@@ -13,9 +13,10 @@ import MeetingMinutes from "@/components/study/CCMeetingMinutes";
 import DownloadButton from "@/components/detail/SCDownloadButton";
 import { formatFileSize } from "@/utils/attachedFileUtils";
 import { getFileIcon } from "@/utils/attachedFileUtils";
-import { calculateDDay, getStatusColor } from "@/utils/studyHelper";
-import { STATUS_LABELS } from "@/types/study";
+import { calculateDDay } from "@/utils/studyHelper";
 import EndRequestButton from "@/components/ui/endRequestButton";
+import { getStatusColor } from "@/utils/badgeUtils";
+import { STATUS_LABELS } from "@/types/progressStatus";
 
 interface StudyDetailPageProps {
   params: { id: string };
@@ -82,7 +83,10 @@ export default async function StudyDetailPage({
                     {studyData.title}
                   </h1>
                   <div className="flex items-center gap-2">
-                    <DefaultBadge className={getStatusColor(studyData.status)}>
+                    <DefaultBadge
+                      variant="custom"
+                      className={getStatusColor(studyData.status)}
+                    >
                       {STATUS_LABELS[studyData.status]}
                     </DefaultBadge>
                     {dDay !== null && (
