+++ src/app/(main)/blog/[id]/page.tsx
@@ -6,7 +6,7 @@ import BackToListButton from "@/components/detail/SCBackToListButton";
 import KebabMenuButton from "@/components/detail/CCKebabMenu";
 import ShareButton from "@/components/detail/CCShareButton";
 import { formatDate } from "@/utils/formatDateUtil";
-import { getCategoryColor } from "@/utils/categoryColorUtils";
+import { getCategoryColor } from "@/utils/badgeUtils";
 import DefaultBadge from "@/components/ui/defaultBadge";
 import CCPublishedCheckbox from "@/components/admin/blog/CCPublishedCheckbox";
 import Link from "next/link";
@@ -73,7 +73,7 @@ export default async function BlogDetailPage({ params }: BlogDetailPageProps) {
           <div className="flex items-start justify-between mb-4 ">
             <div>
               <DefaultBadge
-                variant="outline"
+                variant="custom"
                 className={getCategoryColor(post.category)}
               >
                 {post.category}
