+++ src/app/(main)/study/[id]/page.tsx
@@ -14,9 +14,10 @@ import MeetingMinutes from "@/components/study/CCMeetingMinutes";
 import DownloadButton from "@/components/detail/SCDownloadButton";
 import { formatFileSize } from "@/utils/attachedFileUtils";
 import { getFileIcon } from "@/utils/attachedFileUtils";
-import { calculateDDay, getStatusColor } from "@/utils/studyHelper";
-import { STATUS_LABELS } from "@/types/study";
+import { calculateDDay } from "@/utils/studyHelper";
 import EndRequestButton from "@/components/ui/endRequestButton";
+import { getStatusColor } from "@/utils/badgeUtils";
+import { STATUS_LABELS } from "@/types/progressStatus";
 
 function getStudyDataById(id: string): StudyDetailData | null {
   const parsedId = parseInt(id, 10);
@@ -86,7 +87,10 @@ export default async function StudyMaterialDetailPage({
                     {studyData.title}
                   </h1>
                   <div className="flex items-center gap-2">
-                    <DefaultBadge className={getStatusColor(studyData.status)}>
+                    <DefaultBadge
+                      variant="custom"
+                      className={getStatusColor(studyData.status)}
+                    >
                       {STATUS_LABELS[studyData.status]}
                     </DefaultBadge>
                     {dDay !== null && (
