+++ src/app/globals.css
@@ -267,6 +267,32 @@ body {
   .input-default {
     @apply h-10 w-full rounded-md border border-gray-300 text-gray-900 bg-white px-3 py-2 text-sm focus:border-cert-red focus:outline-none;
   }
+
+  /* Badge 색상 */
+  .badge-gray {
+    @apply bg-gray-50 text-gray-600 border-gray-200;
+  }
+  .badge-blue {
+    @apply bg-blue-50 text-blue-600 border-blue-200;
+  }
+  .badge-green {
+    @apply bg-green-50 text-green-600 border-green-200;
+  }
+  .badge-orange {
+    @apply bg-orange-50 text-orange-600 border-orange-200;
+  }
+  .badge-red {
+    @apply bg-red-50 text-red-600 border-red-200;
+  }
+  .badge-purple {
+    @apply bg-purple-50 text-purple-600 border-purple-200;
+  }
+  .badge-sky {
+    @apply bg-sky-50 text-sky-600 border-sky-200;
+  }
+  .badge-yellow {
+    @apply bg-yellow-50 text-yellow-600 border-yellow-200;
+  }
 }
 
 @keyframes flyAcross {
