+++ src/components/profile/SCBlogList.tsx
@@ -7,7 +7,7 @@ import { ProfileBlogDataType } from "@/types/profile";
 import Link from "next/link";
 import { Plus } from "lucide-react";
 import { BlogCategory } from "@/types/blog";
-import { getCategoryColor } from "@/utils/categoryColorUtils";
+import { getCategoryColor } from "@/utils/badgeUtils";
 
 interface SCBlogListProps {
   searchParams: Promise<{
@@ -54,9 +54,10 @@ export default async function SCBlogList({
                       <div className="flex items-center gap-2 mt-2 text-sm text-gray-600 transition-colors duration-300">
                         <span>{blog.createdAt}</span>
                         <DefaultBadge
-                          className={`border-gray-200 text-gray-600 ml-2 ${getCategoryColor(
+                          variant="custom"
+                          className={getCategoryColor(
                             blog.category as BlogCategory
-                          )}`}
+                          )}
                         >
                           {blog.category}
                         </DefaultBadge>
